(PC-28173)[PRO] refactor: move an rename error helper
