+++ pro/src/apiClient/helpers.ts
@@ -13,18 +13,40 @@ export const serializeApiErrors = (
   return errors
 }
 
+// defaultMessage = 'Une erreur s’est produite, veuillez réessayer'
+
 export const getErrorCode = (error: ApiError): string => {
   return error.body.code
 }
 
-export const isError = (error: any): error is Error =>
-  typeof error === 'object' && 'message' in error
+export const isError = (error: unknown): error is Error =>
+  typeof error === 'object' && error !== null && 'message' in error
 
-export const isErrorAPIError = (error: any): error is ApiError =>
-  'name' in error && error.name === 'ApiError'
+export const isErrorAPIError = (error: unknown): error is ApiError =>
+  isError(error) && 'name' in error && error.name === 'ApiError'
 
 export const getError = (error: ApiError): any => {
   return error.body
+}
+
+export const getHumanReadableApiError = (errors: any) => {
+  if (errors instanceof Array) {
+    return errors
+      .map((error) =>
+        Object.keys(error)
+          .map((key) => error[key])
+          .join(' ')
+      )
+      .join(' ')
+  }
+
+  if (errors instanceof Object) {
+    return Object.keys(errors)
+      .map((key) => errors[key].map((error: any) => error).join(' '))
+      .join(' ')
+  }
+
+  return ''
 }
 
 // FIXME: find a way to test this by mocking ReadableStream
