+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/AllocineProviderParamaters/AllocineProviderParameters.tsx
@@ -1,7 +1,11 @@
 import React, { useState } from 'react'
 
 import { api } from 'apiClient/api'
-import { getError, isErrorAPIError } from 'apiClient/helpers'
+import {
+  getError,
+  getHumanReadableApiError,
+  isErrorAPIError,
+} from 'apiClient/helpers'
 import { PostVenueProviderBody, VenueProviderResponse } from 'apiClient/v1'
 import useNotification from 'hooks/useNotification'
 import { ButtonVariant } from 'ui-kit/Button/types'
@@ -9,7 +13,6 @@ import { Button } from 'ui-kit/index'
 
 import { FormValuesProps } from '../AllocineProviderForm/AllocineProviderForm'
 import AllocineProviderFormDialog from '../AllocineProviderFormDialog/AllocineProviderFormDialog'
-import { getRequestErrorStringFromErrors } from '../utils/getRequestErrorStringFromErrors'
 
 import style from './AllocineProviderParameters.module.scss'
 
@@ -42,7 +45,7 @@ const AllocineProviderParameters = ({
       .catch((error) => {
         isSucess = false
         if (isErrorAPIError(error)) {
-          notification.error(getRequestErrorStringFromErrors(getError(error)))
+          notification.error(getHumanReadableApiError(getError(error)))
         }
       })
