--- pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/utils/getRequestErrorStringFromErrors.ts
@@ -1,19 +1,0 @@
-export const getRequestErrorStringFromErrors = (errors: any) => {
-  if (errors instanceof Array) {
-    return errors
-      .map((error) =>
-        Object.keys(error)
-          .map((key) => error[key])
-          .join(' ')
-      )
-      .join(' ')
-  }
-
-  if (errors instanceof Object) {
-    return Object.keys(errors)
-      .map((key) => errors[key].map((error: any) => error).join(' '))
-      .join(' ')
-  }
-
-  return ''
-}
